# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Styr API
  description: >
    Styr's REST and SSE API: OIDC login, Claude Code sessions, approvals,
    workspaces, profiles and server administration.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: /
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /api/v1/triggers:
    get:
      summary: List triggers
      description: Every trigger visible to the current user (owned, shared, or — for an admin — all).
      operationId: listTriggers
      tags: [triggers]
      responses:
        "200":
          description: Every visible trigger (never includes a secret)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trigger"
    post:
      summary: Create a trigger
      description: >
        Styr generates the bearer secret; it is returned in this response
        only — GET/PATCH/list never include it again, only secret_hint.
        shared marks the trigger as owned by nobody; only an admin may set
        it.
      operationId: createTrigger
      tags: [triggers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  trigger:
                    $ref: "#/components/schemas/Trigger"
                  secret:
                    type: string
                required: [trigger, secret]
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/triggers/samples/{kind}:
    get:
      summary: Get a realistic sample payload for a trigger kind
      description: >
        Used for "send test payload" and template dry-run preview in the
        UI. This literal path is matched before /triggers/{id}, so "samples"
        is never treated as a trigger id. An unrecognized kind is a 404.
      operationId: getTriggerSample
      tags: [triggers]
      parameters:
        - name: kind
          in: path
          required: true
          schema:
            type: string
            enum: [generic, grafana, github]
      responses:
        "200":
          description: A sample webhook body for kind
          content:
            application/json:
              schema:
                type: object
        "404":
          description: Unrecognized kind
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/triggers/{id}:
    get:
      summary: Get a trigger
      operationId: getTrigger
      tags: [triggers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The trigger (never includes a secret)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trigger"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      summary: Update a trigger
      operationId: patchTrigger
      tags: [triggers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerInput"
      responses:
        "200":
          description: The updated trigger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trigger"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a trigger
      operationId: deleteTrigger
      tags: [triggers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/triggers/{id}/rotate-secret:
    post:
      summary: Rotate a trigger's secret
      description: Invalidates the previous secret; the new plaintext secret is returned only in this response.
      operationId: rotateTriggerSecret
      tags: [triggers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The new secret
          content:
            application/json:
              schema:
                type: object
                properties:
                  secret:
                    type: string
                required: [secret]
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/triggers/{id}/test:
    post:
      summary: Run the delivery pipeline against a test payload
      description: >
        Runs the same pipeline as an inbound POST /hooks/{slug} call, as an
        authenticated request: dedupe, cooldown and the storm cap apply
        unless force is true. Responds with the same {delivery_id, status,
        run_id?} shape as POST /hooks/{slug}.
      operationId: testTrigger
      tags: [triggers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payload:
                  type: object
                force:
                  type: boolean
              required: [payload]
      responses:
        "200":
          description: The resulting delivery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryResult"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: styr_session
    bearerToken:
      type: http
      scheme: bearer
      description: >
        A personal API token (POST /me/api-tokens), sent as `Authorization:
        Bearer styr_pat_...`. Equivalent to sessionCookie for every route,
        except that it is exempt from the X-Requested-With CSRF header
        rule (internal/api/middleware.go's csrfGuard).
  responses:
...
    Forbidden:
      description: The current user may not perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: Not found (or not visible to the current user)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    TriggerInput:
      description: The create/update request body for POST/PATCH /triggers(/{id}).
      type: object
      properties:
        name:
          type: string
        kind:
          type: string
          enum: [generic, grafana, github]
        template_id:
          type: string
        dedupe_key_template:
          type: string
        cooldown_s:
          type: integer
        storm_cap_per_hour:
          type: integer
        run_on_resolved:
          type: boolean
        shared:
          type: boolean
          description: Owned by nobody rather than the caller. Admin only.
        enabled:
          type: boolean
          description: Update only; omit to leave enabled state unchanged.
      required: [name, kind, template_id]
    Trigger:
      description: Never includes the secret or its hash — see secret_hint.
      type: object
      properties:
        id:
          type: string
        owner_id:
          type: [string, "null"]
        name:
          type: string
        slug:
          type: string
        kind:
          type: string
          enum: [generic, grafana, github]
        secret_hint:
          type: string
        template_id:
          type: string
        enabled:
          type: boolean
        dedupe_key_template:
          type: string
        cooldown_s:
          type: integer
        storm_cap_per_hour:
          type: integer
        run_on_resolved:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        last_delivery_at:
          type: [string, "null"]
          format: date-time
      required:
        - id
        - owner_id
        - name
        - slug
        - kind
        - secret_hint
        - template_id
        - enabled
        - dedupe_key_template
        - cooldown_s
        - storm_cap_per_hour
        - run_on_resolved
        - created_at
        - updated_at
        - last_delivery_at
...
    DeliveryResult:
      description: >
        The shape POST /hooks/{slug} and POST /triggers/{id}/test both
        respond with: enough to look the delivery up (or follow it to the
        run it started) without the full Delivery record.
      type: object
      properties:
        delivery_id:
          type: string
        status:
          type: string
          enum: [accepted, deduped, cooldown, storm, rejected, failed, skipped]
        run_id:
          type: string
      required: [delivery_id, status]
...